pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-shell"
version = "0.1.0"
description = "A production-ready CLI to chat with OpenAI models, with streaming, config, and history."
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Sowrhoop Raaj B" }]
license = { text = "AGPL-3.0-or-later" }
dependencies = [
  "openai>=1.25.0",
  "typer>=0.12.3",
  "rich>=13.7.1",
  "platformdirs>=4.2.2",
  "python-dotenv>=1.0.1",
  "tenacity>=8.2.3"
]

[project.urls]
Homepage = "https://github.com/sowrhoop/AI-Shell"
Repository = "https://github.com/sowrhoop/AI-Shell"

[project.scripts]
ai-shell = "ai_shell.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_shell*"]

